Improve speed
Don't use minimization automatically any more. Use simply a removal of the unreachable states. This is not the same thing and leads to DFA that are not minimal, but the algorith is linear. Compilation times are reduced dramatically and, in practice, very good results are obtained, that don't affect further operations significantly (but reduce)